+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - `isDirty` validation for Checkbox Element
+- default value for date not catering for timezone
 
 ## [2.13.0] - 2023-04-27